requirements.txt
aniso8601==9.0.1
asgiref==3.3.4
attrs==21.2.0
Babel==2.9.1
base45==0.2.1
certifi~=2021.10.8
cftime==1.5.0
click==8.0.1
cose==0.9.dev8
cycler==0.10.0
Django==3.2.4
django-qrcode==0.3
Flask==2.0.1
Flask-Babel==2.0.0
Flask-RESTful==0.3.9
Flask-SSLify==0.1.5
flynn==1.0.0b2
gunicorn==20.0.4
image==1.5.33
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
numpy==1.21.0
pandas==1.3.4
Pillow==8.4.0
pycryptodome==3.11.0
pyerfa==2.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzbar==0.1.8
scipy==1.7.2
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.1
Werkzeug==2.0.1
xarray==0.15.1